+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [main]
+
+### Changed
+
+- For extension authors: extensions binding to a TCP or UDP port to provide a
+  service should register the port using the `use_port()` context manager if
+  they wish to expose the port number to others. Registered ports will also be
+  made public via the `SYS-PORTS` message.
+
 ## [2.22.0] - 2024-09-18
 
 ### Added